+++ scripts/constants.py
@@ -9,3 +9,7 @@ if "CONFIG_DIR" in os.environ:
 TEST_INTEGRATION_NAME_PREFIX = "test_"
 
 TEST_INTEGRATION_NAME_SUFFIX = "_ignore"
+
+REQUEST_TIMEOUT = 10
+
+HEADER = {"Content-Type": "application/json"}
